@@ -2,31 +2,33 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
 using System.Windows.Forms;
-
-using CreamInstaller.Classes;
-using CreamInstaller.Forms.Components;
-
+using CreamInstaller.Components;
+using CreamInstaller.Utility;
 using HtmlAgilityPack;
-
 using Onova;
 using Onova.Models;
 using Onova.Services;
 
-namespace CreamInstaller;
+namespace CreamInstaller.Forms;
 
-internal partial class MainForm : CustomForm
+internal sealed partial class MainForm : CustomForm
 {
-    internal MainForm() : base()
+    private CancellationTokenSource cancellationTokenSource;
+    private Version latestVersion;
+
+    private UpdateManager updateManager;
+    private IReadOnlyList<Version> versions;
+
+    internal MainForm()
     {
         InitializeComponent();
-        Text = Program.ApplicationName;
+        Text = Program.ApplicationNameShort;
     }
-
-    private static CancellationTokenSource cancellationTokenSource;
 
     private void StartProgram()
     {
@@ -36,87 +38,111 @@ internal partial class MainForm : CustomForm
             cancellationTokenSource.Dispose();
             cancellationTokenSource = null;
         }
+#pragma warning disable CA2000 // Dispose objects before losing scope
+        SelectForm form = new();
+#pragma warning restore CA2000 // Dispose objects before losing scope
+        form.InheritLocation(this);
+        form.FormClosing += (_, _) => Close();
+        form.Show();
         Hide();
-        new SelectForm(this).ShowDialog();
-        Close();
+#if DEBUG
+        DebugForm.Current.Attach(form);
+#endif
     }
-
-    private static UpdateManager updateManager = null;
-    private static Version latestVersion = null;
-    private static IReadOnlyList<Version> versions;
 
     private async void OnLoad()
     {
-        Size = new(420, 85);
-        progressBar1.Visible = false;
+        progressBar.Visible = false;
         ignoreButton.Visible = true;
         updateButton.Text = "Update";
         updateButton.Click -= OnUpdateCancel;
-        label1.Text = "Checking for updates . . .";
+        progressLabel.Text = "Checking for updates . . .";
         changelogTreeView.Visible = false;
-        changelogTreeView.Location = new(12, 41);
-        changelogTreeView.Size = new(380, 208);
-
-        GithubPackageResolver resolver = new("pointfeev", "CreamInstaller", "CreamInstaller.zip");
+        changelogTreeView.Location = progressLabel.Location with { Y = progressLabel.Location.Y + progressLabel.Size.Height + 13 };
+        Refresh();
+#if DEBUG
+        DebugForm.Current.Attach(this);
+#endif
+        GithubPackageResolver resolver = new(Program.RepositoryOwner, Program.RepositoryName, Program.RepositoryPackage);
         ZipPackageExtractor extractor = new();
         updateManager = new(AssemblyMetadata.FromAssembly(Program.EntryAssembly, Program.CurrentProcessFilePath), resolver, extractor);
-
         if (latestVersion is null)
         {
-            CheckForUpdatesResult checkForUpdatesResult = null;
             cancellationTokenSource = new();
             try
             {
-                checkForUpdatesResult = await updateManager.CheckForUpdatesAsync(cancellationTokenSource.Token);
-                cancellationTokenSource.Dispose();
-                cancellationTokenSource = null;
+                CheckForUpdatesResult checkForUpdatesResult = await updateManager.CheckForUpdatesAsync(cancellationTokenSource.Token);
+#if !DEBUG
                 if (checkForUpdatesResult.CanUpdate)
                 {
-                    latestVersion = checkForUpdatesResult.LastVersion;
-                    versions = checkForUpdatesResult.Versions;
+#endif
+                latestVersion = checkForUpdatesResult.LastVersion;
+                versions = checkForUpdatesResult.Versions;
+#if !DEBUG
                 }
+#endif
             }
-            catch { }
+#if DEBUG
+            catch (TaskCanceledException) { }
+            catch (Exception e)
+            {
+                DebugForm.Current.Log($"Exception while checking for updates: {e.GetType()} ({e.Message})", LogTextBox.Warning);
+            }
+#else
+            catch
+            {
+                // ignored
+            }
+#endif
+            finally
+            {
+                cancellationTokenSource.Dispose();
+                cancellationTokenSource = null;
+            }
         }
-
         if (latestVersion is null)
         {
             updateManager.Dispose();
+            updateManager = null;
             StartProgram();
         }
         else
         {
-            Size = new(420, 300);
-            label1.Text = $"An update is available: v{latestVersion}";
+            progressLabel.Text = $"An update is available: v{latestVersion}";
             ignoreButton.Enabled = true;
             updateButton.Enabled = true;
-            updateButton.Click += new(OnUpdate);
+            updateButton.Click += OnUpdate;
             changelogTreeView.Visible = true;
-            Version currentVersion = new(Application.ProductVersion);
-            foreach (Version version in versions)
-                if (version > currentVersion && !changelogTreeView.Nodes.ContainsKey(version.ToString()))
+            Version currentVersion = new(Program.Version);
+#if DEBUG
+            foreach (Version version in versions.Where(v => (v > currentVersion || v == latestVersion) && !changelogTreeView.Nodes.ContainsKey(v.ToString())))
+#else
+            foreach (Version version in versions.Where(v => v > currentVersion && !changelogTreeView.Nodes.ContainsKey(v.ToString())))
+#endif
+            {
+                TreeNode root = new($"v{version}") { Name = version.ToString() };
+                changelogTreeView.Nodes.Add(root);
+                if (changelogTreeView.Nodes.Count > 0)
+                    changelogTreeView.Nodes[0].EnsureVisible();
+                _ = Task.Run(async () =>
                 {
-                    TreeNode root = new($"v{version}");
-                    root.Name = root.Text;
-                    changelogTreeView.Nodes.Add(root);
-                    _ = Task.Run(async () =>
-                    {
-                        HtmlNodeCollection nodes = await HttpClientManager.GetDocumentNodes(
-                            $"https://github.com/pointfeev/CreamInstaller/releases/tag/v{version}",
-                            "//div[@data-test-selector='body-content']/ul/li");
-                        if (nodes is null) changelogTreeView.Nodes.Remove(root);
-                        else foreach (HtmlNode node in nodes)
+                    HtmlNodeCollection nodes = await HttpClientManager.GetDocumentNodes(
+                        $"https://github.com/{Program.RepositoryOwner}/{Program.RepositoryName}/releases/tag/v{version}",
+                        "//div[@data-test-selector='body-content']/ul/li");
+                    if (nodes is null)
+                        changelogTreeView.Nodes.Remove(root);
+                    else
+                        foreach (HtmlNode node in nodes)
+                            changelogTreeView.Invoke(delegate
                             {
-                                Program.Invoke(changelogTreeView, delegate
-                                {
-                                    TreeNode change = new();
-                                    change.Text = HttpUtility.HtmlDecode(node.InnerText);
-                                    root.Nodes.Add(change);
-                                    root.Expand();
-                                });
-                            }
-                    });
-                }
+                                TreeNode change = new() { Text = HttpUtility.HtmlDecode(node.InnerText) ?? string.Empty };
+                                root.Nodes.Add(change);
+                                root.Expand();
+                                if (changelogTreeView.Nodes.Count > 0)
+                                    changelogTreeView.Nodes[0].EnsureVisible();
+                            });
+                });
+            }
         }
     }
 
@@ -125,21 +151,24 @@ internal partial class MainForm : CustomForm
     retry:
         try
         {
-            string FileName = Path.GetFileName(Program.CurrentProcessFilePath);
-            if (FileName != "CreamInstaller.exe")
-                if (new DialogForm(this).Show(Program.ApplicationName, SystemIcons.Warning,
-                    "WARNING: CreamInstaller.exe was renamed!" +
-                    "\n\nThis will cause unwanted behavior when updating the program!",
-                    "Ignore", "Abort") == DialogResult.Cancel)
+            string fileName = Path.GetFileName(Program.CurrentProcessFilePath);
+            if (fileName != Program.ApplicationExecutable)
+            {
+                using DialogForm form = new(this);
+                if (form.Show(SystemIcons.Warning,
+                        "WARNING: " + Program.ApplicationExecutable + " was renamed!" + "\n\nThis will cause undesirable behavior when updating the program!",
+                        "Ignore", "Abort") == DialogResult.Cancel)
                 {
                     Application.Exit();
                     return;
                 }
+            }
             OnLoad();
         }
         catch (Exception e)
         {
-            if (ExceptionHandler.OutputException(e)) goto retry;
+            if (e.HandleException(this))
+                goto retry;
             Close();
         }
     }
@@ -148,44 +177,64 @@ internal partial class MainForm : CustomForm
 
     private async void OnUpdate(object sender, EventArgs e)
     {
-        progressBar1.Visible = true;
+        progressBar.Visible = true;
         ignoreButton.Visible = false;
         updateButton.Text = "Cancel";
         updateButton.Click -= OnUpdate;
-        updateButton.Click += new(OnUpdateCancel);
-        changelogTreeView.Location = new(12, 70);
-        changelogTreeView.Size = new(380, 179);
-
+        updateButton.Click += OnUpdateCancel;
+        changelogTreeView.Location = progressBar.Location with { Y = progressBar.Location.Y + progressBar.Size.Height + 6 };
+        Refresh();
         Progress<double> progress = new();
-        progress.ProgressChanged += new(delegate (object sender, double _progress)
+        progress.ProgressChanged += delegate(object _, double _progress)
         {
-            label1.Text = $"Updating . . . {(int)_progress}%";
-            progressBar1.Value = (int)_progress;
-        });
-
-        label1.Text = "Updating . . . ";
+            progressLabel.Text = $"Updating . . . {(int)_progress}%";
+            progressBar.Value = (int)_progress;
+        };
+        progressLabel.Text = "Updating . . . ";
         cancellationTokenSource = new();
         try
         {
             await updateManager.PrepareUpdateAsync(latestVersion, progress, cancellationTokenSource.Token);
+        }
+#if DEBUG
+        catch (TaskCanceledException) { }
+        catch (Exception ex)
+        {
+            DebugForm.Current.Log($"Exception while preparing update: {ex.GetType()} ({ex.Message})", LogTextBox.Warning);
+        }
+#else
+        catch
+        {
+            // ignored
+        }
+#endif
+        finally
+        {
             cancellationTokenSource.Dispose();
             cancellationTokenSource = null;
         }
-        catch { }
-
         if (updateManager is not null && updateManager.IsUpdatePrepared(latestVersion))
         {
             updateManager.LaunchUpdater(latestVersion);
             Application.Exit();
             return;
         }
-        else OnLoad();
+        OnLoad();
     }
 
     private void OnUpdateCancel(object sender, EventArgs e)
     {
-        cancellationTokenSource.Cancel();
-        updateManager.Dispose();
+        cancellationTokenSource?.Cancel();
+        updateManager?.Dispose();
         updateManager = null;
+    }
+
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+            components?.Dispose();
+        base.Dispose(disposing);
+        cancellationTokenSource?.Dispose();
+        updateManager?.Dispose();
     }
 }
