@@ -6,84 +6,96 @@ using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Windows.Forms;
-
-using CreamInstaller.Classes;
+using CreamInstaller.Forms;
+using CreamInstaller.Platforms.Steam;
+using CreamInstaller.Utility;
 
 namespace CreamInstaller;
 
 internal static class Program
 {
-    internal static readonly string ApplicationName = Application.CompanyName + " v" + Application.ProductVersion + ": " + Application.ProductName;
+    internal static readonly string Name = Application.CompanyName;
+    private static readonly string Description = Application.ProductName;
+    internal static readonly string Version = Application.ProductVersion;
+
+    internal const string RepositoryOwner = "pointfeev";
+    internal static readonly string RepositoryName = Name;
+    internal static readonly string RepositoryPackage = Name + ".zip";
+#if DEBUG
+    internal static readonly string ApplicationName = Name + " v" + Version + "-debug: " + Description;
+    internal static readonly string ApplicationNameShort = Name + " v" + Version + "-debug";
+    internal static readonly string ApplicationExecutable = Name + "-debug.exe"; // should be the same as in .csproj
+#else
+    internal static readonly string ApplicationName = Name + " v" + Version + ": " + Description;
+    internal static readonly string ApplicationNameShort = Name + " v" + Version;
+    internal static readonly string ApplicationExecutable = Name + ".exe"; // should be the same as in .csproj
+#endif
 
     internal static readonly Assembly EntryAssembly = Assembly.GetEntryAssembly();
-    internal static readonly Process CurrentProcess = Process.GetCurrentProcess();
-    internal static readonly string CurrentProcessFilePath = CurrentProcess.MainModule.FileName;
+    private static readonly Process CurrentProcess = Process.GetCurrentProcess();
+    internal static readonly string CurrentProcessFilePath = CurrentProcess.MainModule?.FileName;
 
     internal static bool BlockProtectedGames = true;
-    internal static readonly string[] ProtectedGameNames = { "PAYDAY 2", "Call to Arms" }; // non-functioning CreamAPI or DLL detections
-    internal static readonly string[] ProtectedGameDirectories = { @"\EasyAntiCheat", @"\BattlEye" }; // DLL detections
-    internal static readonly string[] ProtectedGameDirectoryExceptions = { "Arma 3" }; // Arma 3's BattlEye doesn't detect DLL changes?
+    internal static readonly string[] ProtectedGames = { "PAYDAY 2" };
+    internal static readonly string[] ProtectedGameDirectories = { @"\EasyAntiCheat", @"\BattlEye" };
+    internal static readonly string[] ProtectedGameDirectoryExceptions = Array.Empty<string>();
 
-    internal static bool IsGameBlocked(string name, string directory)
+    internal static bool IsGameBlocked(string name, string directory = null)
     {
-        if (!BlockProtectedGames) return false;
-        if (ProtectedGameNames.Contains(name)) return true;
-        if (!ProtectedGameDirectoryExceptions.Contains(name))
-            foreach (string path in ProtectedGameDirectories)
-                if (Directory.Exists(directory + path)) return true;
-        return false;
-    }
-
-    internal static bool IsFilePathLocked(this string filePath)
-    {
-        try { File.Open(filePath, FileMode.Open, FileAccess.ReadWrite, FileShare.None).Close(); }
-        catch (FileNotFoundException) { return false; }
-        catch (IOException) { return true; }
-        return false;
+        if (!BlockProtectedGames)
+            return false;
+        if (ProtectedGames.Contains(name))
+            return true;
+        if (directory is null || ProtectedGameDirectoryExceptions.Contains(name))
+            return false;
+        return ProtectedGameDirectories.Any(path => Directory.Exists(directory + path));
     }
 
     internal static bool IsProgramRunningDialog(Form form, ProgramSelection selection)
     {
-        if (selection.AreSteamApiDllsLocked)
+        while (true)
         {
-            if (new DialogForm(form).Show(ApplicationName, SystemIcons.Error,
-            $"ERROR: {selection.Name} is currently running!" +
-            "\n\nPlease close the program/game to continue . . . ",
-            "Retry", "Cancel") == DialogResult.OK)
-                return IsProgramRunningDialog(form, selection);
+            if (selection.AreDllsLocked)
+            {
+                using DialogForm dialogForm = new(form);
+                if (dialogForm.Show(SystemIcons.Error,
+                        $"ERROR: {selection.Name} is currently running!" + "\n\nPlease close the program/game to continue . . . ", "Retry", "Cancel")
+                 == DialogResult.OK)
+                    continue;
+            }
+            else
+                return true;
+            return false;
         }
-        else return true;
-        return false;
-    }
-
-    internal static void GetApiComponents(this string directory, out string api, out string api_o, out string api64, out string api64_o, out string cApi)
-    {
-        api = directory + @"\steam_api.dll";
-        api_o = directory + @"\steam_api_o.dll";
-        api64 = directory + @"\steam_api64.dll";
-        api64_o = directory + @"\steam_api64_o.dll";
-        cApi = directory + @"\cream_api.ini";
     }
 
     [STAThread]
     private static void Main()
     {
-        Mutex mutex = new(true, "CreamInstaller", out bool createdNew);
+        using Mutex mutex = new(true, Name, out bool createdNew);
         if (createdNew)
         {
-            Application.SetHighDpiMode(HighDpiMode.SystemAware);
+            _ = Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.ApplicationExit += new(OnApplicationExit);
+            Application.ApplicationExit += OnApplicationExit;
+            Application.ThreadException += (_, e) => e.Exception.HandleFatalException();
+            Application.SetUnhandledExceptionMode(UnhandledExceptionMode.CatchException);
+            AppDomain.CurrentDomain.UnhandledException += (_, e) => (e.ExceptionObject as Exception)?.HandleFatalException();
         retry:
             try
             {
                 HttpClientManager.Setup();
-                Application.Run(new MainForm());
+                using MainForm form = new();
+#if DEBUG
+                DebugForm.Current.Attach(form);
+#endif
+                Application.Run(form);
             }
             catch (Exception e)
             {
-                if (ExceptionHandler.OutputException(e)) goto retry;
+                if (e.HandleException())
+                    goto retry;
                 Application.Exit();
                 return;
             }
@@ -91,9 +103,8 @@ internal static class Program
         mutex.Close();
     }
 
-    internal static void Invoke(this Control control, MethodInvoker methodInvoker) => control.Invoke(methodInvoker);
+    internal static bool Canceled;
 
-    internal static bool Canceled = false;
     internal static async void Cleanup(bool cancel = true)
     {
         Canceled = cancel;
